Refractor functions into files and remove await at top level of script #JC19-353
