+++ apps/vanilla/src/pages/index.html
@@ -3,26 +3,28 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- Compiled and minified CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
     <link rel="stylesheet" href="./index.css" />
     <title>Vite App</title>
   </head>
-  <body class="default-layout">
-    <h1 class="title">Saritasa Anime</h1>
-    <div class="container">
+  <body class="page default-layout">
+    <h1 class="page__title">Saritasa Anime</h1>
+    <div class="page__content">
       <!-- Filter -->
       <div class="container__filter">
+
         <div class="filter__item">
-          <h2 class="select__title">Sort by</h2>
+          <label class="select__title">Sort by</label>
           <select class="select select__sort">
           </select>
         </div>
+
         <div class="filter__item">
-          <h2 class="select__title">Ordering by</h2>
+          <label class="select__title">Ordering by</label>
           <select class="select select__order">
           </select>
         </div>
+
       </div>
       <!-- Anime films list -->
       <div class="container__table">
